farada/requirements.txt
Django==2.2.5
django-crispy-forms==1.7.2
django-multiselectfield==0.1.8
mysqlclient==1.3.12
Pillow==6.1.0
pytz==2019.2
sqlparse==0.3.0
